abstract common logic into reusable hook
