+++ client/landing/stepper/hooks/use-send-email-verification.ts
@@ -1,9 +1,7 @@
 import wpcom from 'calypso/lib/wp';
 
 export function useSendEmailVerification() {
-	const sendEmailVerification = async () => {
+	return async () => {
 		return wpcom.req.post( '/me/send-verification-email', { apiVersion: '1.1' } );
 	};
-
-	return sendEmailVerification;
 }
